[parser] remove redundant error when using enum keyword in function signature
Summary:
previously, either of these would result in 2 errors:

```
function enum() { "use strict" }
function foo(enum) { "use strict" }
```

one because "enum" is a reserved word, and another complaining that enum is a reserved word in strict mode.

the latter is redundant, because it's always reserved, not just in strict mode.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D42606007

fbshipit-source-id: 1a13528cf3d0dad28a59e3089fa3e3b9ab8e0f7b