[parser] add missing reserved types to is_reserved_type
Summary:
we added `keyof`, `never`, `undefined` and `unknown` as reserved types, but didn't add them to the list of reserved types!

`type never = ...` should error instead of silently ignoring the alias because uses of `never` are parsed as `T_NEVER` and so the `never` type in the env is never looked up in practice.

Changelog: [internal] already covered by the commits reserving each of these types

Reviewed By: gkz

Differential Revision: D42597843

fbshipit-source-id: 8769a078977df5a204855b25027ced742258881a