[parser] fix regression banning async in for await loop
Summary:
```
for (async of [7]);
```
this is invalid because it can't be parsed in LR(2): you can't tell if it's a for loop with an async arrow function with an `of` param, or a for-of loop with an `async` identifier.

however:

```
for await (async of [7]);
```
this is valid, because a for-await loop isn't ambiguous.

this regressed in D42187354 (https://github.com/facebook/flow/commit/5c05dbe6b625d8ca133e2fe4d4a82024fc1be29a)

Fixes https://github.com/facebook/flow/issues/8987

Changelog: [parser] Fix parsing of `for await (async of ...)` (regressed in 0.197.0)

Reviewed By: gkz

Differential Revision: D42607157

fbshipit-source-id: b76f11eedb044436eaf7c08f57590be81eca4d49