[generic-escape] Predicate function body won't have escaped generics
Summary:
As soon as we finished the new env implementation, it should be clear the body of the predicate functions can never cause generic types to escape, because the predicate type is produced from refinement data structures, which is completely syntactic and it knows nothing about the generic types.

Changelog: [fixed] We will no longer emit escaped-generic errors for predicate function bodys.

Reviewed By: mvitousek

Differential Revision: D42594315

fbshipit-source-id: 5e92425ce92078f0b27aa52342bc5776ea129966