Drop support for spreading in compose
Summary:
In the next diff where I tried to fix $Compose under LTI, I found that $Compose with spreads is still not well behaving.

It is also not very useful. Conceptually, it's only doing to give you useful results when all the functions in an array have the same type like `A=>A`. In this case, you can easily do that in userland.

Therefore, I decide to drop the support.

Changelog: [errors] Support for spread argument in the builtin special-cased $Compose function is dropped. If you need this, you can write it yourself:

```
declare function compose<T>(
  fns:...$ReadOnlyArray<T=>T>,
): T;
```

Reviewed By: mvitousek

Differential Revision: D42595204

fbshipit-source-id: c136c4895061649e38e427e80a21b3c4535d3fef