Don't hint array elements after a spread
Summary:
Currently we apply per-index hints to each element of an array literal, but the index of the element can be inconsistent with its index in the resulting array if the element occurs after a spread. This can result in incorrect behavior in cases like

```
declare function arb<T>(): T;
const z: [string, number, number] = ...
const w: [string, number, number, number] = [...z, ...arb()];
```
where we'll report an error that the RHS of `w` has 6 elements--because the hint on `arb()` was an array [number, number, number].

This diff changes the hints after a spread has occurred, since it's impossible to know what element in the hint type corresponds to what element in the literal at that point. For non-spread elements, we still know that they must be some element of the hint, so we can change the hint op to perform a non-indexed `GetElemT` on the hint. For subsequent spreads, there isn't an analogous operation we can perform, so we simply remove the hints.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D42609744

fbshipit-source-id: 3f59ab6f6c5ac61560174d22bf70d77846c84c39