refactor: flatten out PolyT cases in Flow_js
Summary:
I have found the nested pattern matching in the PolyT case rather impractical
in quickly detecting cases where polymorphic types are handled specially.

This diff flattens out these cases of the pattern match so that they are all of the
form `PolyT _ ~> SomeUseCtorT _`.

This makes it easier to apply the fix in the next diff.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D42641961

fbshipit-source-id: bad83cd1d8703bb7b57edc6b3439119e62749b5b