Connect .js and .js.flow files
Summary:
Recall that a file module can be provided by either a .js.flow or a .js file, in
that order.

Before this diff, to determine the provider, we constructed a .js.flow key and
checked the hash table for that key before falling back to the .js file.

Instead, this diff adds a pointer from the .js file to the .js.flow file. To get
the provider, we look up the .js file, then check for a linked .js.flow file.

In almost all cases, the link will be null, so we avoid some work: we don't need
to construct a File_key.t (string concat), hash that key, probe the hash table.

This will be particularly useful in combination with D42161883 (https://github.com/facebook/flow/commit/6a12916ecb985820373a11ed428935543d8cc77e), where we store
resolved modules as addrs instead of Modulename.t values. When resolved modules
are stored as addrs, we would like to find the provider for file module
dependencies directly -- by dereferencing the addr, which this change enables.

The implementation is pretty straightforward. When we create a .js.flow file, we
will either find an existing .js file or create a phantom record to store the
dependents. In other words, we will always have a .js file if we have a .js.flow
file. This is why we only need to add a link when creating a .js.flow file.

Reviewed By: jbrown215

Differential Revision: D42225457

fbshipit-source-id: d5a20398268e19537bd68f972ac9c1856e8ff83f