fix crash on Windows
Summary:
since `MARK_STACK_INIT_SIZE = 512` which is a power of 2, and we resize the mark stack when it's a power of 2, we need to resize the mark stack when `mark_ptr == 512`.

this meant that we were writing to unreserved pages on Windows, leading to access violations (0xC0000005), whenever the mark stack got bigger than 512 entries.

Changelog: [fix] Fix a crash on Windows

Reviewed By: samwgoldman

Differential Revision: D42712031

fbshipit-source-id: 305263be3bb29c5244d584d3642d62e23847c9c8