Fix naked pointer in fsevents code
Summary:
When using dfind, Flow uses fsevents on Mac to get file system notifications.
The fsevents C interop code used naked pointers -- returning a pointer outside
of OCaml's managed heap to OCaml as a `value` -- which is not supported in OCaml 5.

This diff wraps the pointer in an "abstract" value, which is a valid OCaml
value, on OCaml's managed heap, which is not scanned.

I could have used a custom value and even attached a finalizer, but we already
leaked these allocations, so this change is simpler.

I could also have likely "just" tagged the pointer, but I didn't want to have to
think about whether the allocation is 2-aligned and how portable that might be.
If this code were hotter, I'd have considered it.

Reviewed By: mroch

Differential Revision: D42754700

fbshipit-source-id: 62463d988b5ba5ea728a3dcaf4962290a1d9a05e