Fix sklist_remove write barrier
Summary:
In D42747270 (https://github.com/facebook/flow/commit/41b7ff81c9dbab47784d3158e1fd6d533a5a5320) I added the write barrier in the wrong place. The skip list node
stores links at multiple levels. The lowest level is a sorted linked list and
the higher levels are "fast lanes" that skip over the list. The higher the lane,
the more nodes it skips.

I only added the write barrier for the code that marks the higher level links,
but we _only_ need it on the lowest level link. The higher level links are by
definition redundant. A node becomes unreachable when the lowest level link is
marked.

Reviewed By: mroch

Differential Revision: D42767383

fbshipit-source-id: ca1eca3583dfbfc8371a7b72aa547018e0668766