[docs] Correctly state `react.runtime` default value
Summary:
https://github.com/facebook/flow/blob/009fb4a849f680266dfa908036070de35669a02a/src/commands/config/flowConfig.ml#L238

We can also change the config to align with our docs, but updating the docs is easier, and it's also the same as TS defaults.

Close https://github.com/facebook/flow/issues/8513.

Changelog: [docs] We have updated the docs to correctly state that, the default value of `react.runtime` option is, in fact, `classic`.

Reviewed By: mroch

Differential Revision: D42778812

fbshipit-source-id: ecdad27a5004c2a5e858de283c5352f272e1617f