[lti] Fix spurious illegal-this error
Summary:
This bug has existed since contextual typing for return statement was introduced. However, since contextual typing wasn't on for a long time, we never discovered it, and the bug is hard to trigger and manifest in observable errors, but now we have one:

```
class Foo {
  returnThis(): this {
    return _ => {};
  }
}
```

In this example, when we are contextually typing `_`, we will evaluate the return annotation as base hint. Since we have not correctly given it a tparams map, `this` will be evaluated out of correct tparam context, and we will get illegal-this error.

Changelog: [fix] Remove spurious `illegal-this` error when a this annotation is used in contextual typing.

Reviewed By: panagosg7

Differential Revision: D42789952

fbshipit-source-id: 17152737a1b8c2d467c31a89b453ad67ab3f580a