fix exceptions on invalid setters
Summary:
setters must have exactly one param. anything else is a parse error.

during typechecking, we raise an exception if this invariant doesn't hold. that's not a safe assumption because we typecheck invalid ASTs for the IDE.

this meant that any file with an invalid setter lost all IDE services until the parse error was fixed.

now we treat a missing param as `any`, and ignore more than one param

Changelog: [fix] Fix IDE services that stopped working while a file contained a setter with the wrong number of parameters.

Reviewed By: SamChou19815

Differential Revision: D42804017

fbshipit-source-id: 445d5365e9421ac0348ab70d9c559f0143ee1176