Context.cache_snapshot
Summary:
When we roll back error sets, typically when in synthesis mode, we need to make sure
we also roll back a number of caches. The caches are there to avoid redoing computations
when the result is known based on inputs. Side-effects, like errors being raised, are also
only performed once. When we clear an error E but fail to clear the cached constraint C that
led to it, then if we hit C again we will fail to properly raise E.

This diff aims at making it easier to restore the various caches we maintain in Context, by
adding a `cache_snapshot` type. In this diff this is only used within Context, but in the
following diffs it will be used in other places where we call `Context.reset_errors`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D42779204

fbshipit-source-id: 233d08a2494bf5eee92e0840c9d528d0b471e09b