Fix ordering bug around synthesizable objects
Summary:
We weren't properly declaring dependencies on member expressions contained within synthesizable objects, because we special-cased the code path for them and didn't go through #expression. Here, we still keep the special casing (because we don't want to necessarily declare all the dependencies on the _object that we normally would) but we make sure to add a dependency on the member expression's writes.

Changelog: [internal]

Closes #8998.

Reviewed By: SamChou19815

Differential Revision: D42816098

fbshipit-source-id: 197308c00b1b4b45a1713a4832bccbb0ef088d08