remove unnecessary Abnormal.Stmts exceptions
Summary:
`Toplevels.toplevels stmts` visits all of the `stmts`, which may throw `Abnormal` exceptions for `throw`, `return`, `break` or `continue`. those exceptions are caught, and the earliest one is re-thrown on the last line of `toplevels`.

then every caller of `toplevels` immediately catches that exception.

instead, just return an `Abnormal.t option`. this is way less confusing and also forces all future callers of `toplevels` to deal with the potential abnormals.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D42811591

fbshipit-source-id: 77414637462da1b039dae9127798dbc9c76bf534