[lti] Fix bug where an function expression is incorrectly cached and never checked again
Summary:
As a performance optimization, we will cache the typed-ast of a synthesized expression, if we decide that the result will be an "acceptable" checking result. We use the criteria of "no-placeholder generated during synthesis" for "acceptable". This is reasonable, because otherwise we would leak placeholders. Under this condition, we should be getting exactly what we will get if we check the same expression without hint.

This scheme, however, still depends on the synthesis mode to be honest of whether it produced any placeholder-like stuff. In the following case, we do found an edge case:

```
declare function foo<V>(v: V, f: (V) => void): void;

foo(
  { elements: [] },
  (value: { elements: Array<string> }) => { /* body */ },
)
```

We encountered a fully annotated function, but the synthesis will still skip its body. While there is nothing wrong with that, we skip the body using the error AST mapper, which doesn't produce placeholders but just regular any. Then at the point when we decide whether to cache, we see that no placeholders are produced, and we will happily cache it. As a result, all possible type errors within the function body will go undetected!

This diff fixes the issue by constructing the skipped body by a placeholder type generating visitor. We can optimize further by detecting whether the function parameters are fully annotated, and if so, check the body, but that will require changes in multiple parts (name-def-ordering, refactoring func_sig so that we keep the return type unresoved, etc). There is theoretical performance gain from this more advanced approach in cases like `poly_call(id((f: Foo) => {...}))`, but most of the lambdas are likely not deeply nested in function calls.

Changelog: [errors] We fixed a bug under LTI that causes us to incorrectly skipping checking for some function expressions [Example](https://flow.org/try/#0CYUwxgNghgTiAEAzArgOzAFwJYHtVJxwB4A1APgAoA3ALnhIBok6KSBKeAXjPipy2Bs6fAQG4AUOMSEK4+PADe8EBBABbEKgwBnOgG0AuvAC+DOfGpQIyEHSUr1mnXQCCMGFACeRbRhhZUAHMeYw5uRQsqKxs6dQAHDE82URMmAHo0+FQcAHdldxwYbXE2cSA). Some previously hidden errors might be revealed.

Reviewed By: panagosg7

Differential Revision: D42824860

fbshipit-source-id: 09d2e496aabd96df3899af90d17d9764f474d231