[lti] Fix missing placeholder propagation in `ElemT`
Summary:
One important soundness guarantee of LTI is achieved by never leaking placeholders into checking mode. Placeholders are produced during synthesis of polymorphic function call when we are evaluating the hints, and they are filtered away at the end of hint eval. Therefore, placeholder types can never leak into checking.

However, the *effect* of placeholders can still leak, as the following example shows:

```
declare function id<T>(T): T;
id([(item) => 1]); // no error :(
```

During the `_ ~> ElemT(AnyT, _)` step of the constrain, we do not propagate the source of the any, where the placeholder information lives. As a result, the tout contains a regular `AnyT.untyped`. Intuitively, this means that the effect of the placeholder is remembered, but the information that there is a placeholder is lost, since the following steps cannot distinguish between the following two states:

1. The any is originated from a placeholder base type.
2. The any originates from an annotated any or error any.

Therefore, we must do placeholder propagation and do it well to guarantee soundness. In this diff, I fixed the specific issue from this example, and added tests to ensure all other hint decompositions won't suffer the same problem. It's found that we are at least fixing two other errors, which I will investigate further.

Changelog: [errors] TODO

Reviewed By: panagosg7

Differential Revision: D42849625

fbshipit-source-id: 466f928bd51e59ad142ed11247cb798b240b5f55