Consistently ignore `%checks` in predicate functions when parameters are invalid
Summary:
When you write something like:

```
export function f({a: b}): boolean %checks {
  return typeof b === 'string';
}
```

Flow will complain that destructuring parameter for predicate function is unsupported, and the predicate-ness of the function will be ignored. However, the latter part is not exactly true. We only really ignore it in one place: we will generate an ordinary function kind. Later, we still use a simple check like `Option.is_some predicate` to decide whether we should treat this function as a predicate.

In statement.ml when computing the signature of the function, the example's return type will be `OpenT([], [boolean])`. In env_resolution, the type will be given a lower bound, so that it becomes `OpenT([OpenPred], [boolean])`. This happens because in both places, it does a simple `Option.is_some predicate` check, so it still believes this function is predicate. However, during `mk_func_sig`, this function will be considered as ordinary, due to the check that will emit the error if necessary, and this fact is recorded by setting the scope kind to ordinary. Then when we are checking the return statement, the return type will be a boolean instead of an `OpenPred`. Thus, in the end, we will have:

```
boolean ~> OpenT([OpenPred], [boolean])
```

This has never become an issue with visible error diff before LTI, because tvars are not resolved. But when tvars are resolved, the flow will become

```
boolean ~> OpenPred
```

, which causes a type error.

I fixed the issue by consistently checking the validity of parameters, so that we will have a consistent view everywhere. It's not possible to not ignore the predicate even with the error, because it will break sanity checks like:

https://github.com/facebook/flow/blob/a66bcb7ae7708921c1773db97815e8cd2e6ab3e9/src/typing/flow_js.ml#L2354-L2355

Changelog: [fixed] Under LTI mode, we will no longer emit spurious incompatibility errors in invalid predicate function like `function f({a: b}): boolean %checks { return typeof b === 'string'; }`. Instead, you will only get error on this unsupported syntax.

Reviewed By: panagosg7

Differential Revision:
D42867837 (https://github.com/facebook/flow/commit/1eebc65c77118b6270566bea6261374d65d1718f)

------------------------------------------------------------------------
(from 819ecf9fa04d52ed636cb4bd8f6153fa586bd37f)

fbshipit-source-id: 19f940dd7ca83d05ddaf9a0fa04bc4bb4a093684