[lti] Move empty array hint eval code into type_hint.ml
Summary:
We are currently doing the final step of hint decomposition for getting the element of empty array outside of type_hint for no good reasons. Currently, the step is missing the guard to ensure that the inferred element type is not an empty tvar, which is causing some soundness issues downstream.

Instead of fixing this specific issue in statement.ml, I just moved the code into type_hint, which already guards against this. This change also makes the code less fragile by having similar implementations in a few places.

Changelog: [errors] We fixed a bug that causes some empty arrays to be incorrectly inferred as `Array<empty>` under LTI. e.g. `Array.from(nullableStringArray ?? [])`. As a result, previously hidden errors might be revealed.

Reviewed By: panagosg7

Differential Revision: D42904866

fbshipit-source-id: a7f25bbd6ddfac98019a68a71106a03009a39b0e