remove unused root path from Fuzzy_path
Summary:
`match_result_t` had a `score_based_root_path` field that was not exposed to ocaml. it is used to give preference to results that are closer to the given `root_path`, when this library is used for matching file paths. since we don't do that, it's dead code.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D42890351

fbshipit-source-id: 7b89c3d0495b9bf1ef0d579f521b9f21b83e2f11