fix missing errors when a file is deleted before init in lazy mode
Summary:
when a file is deleted, we need to dirty the haste module it provides so that a new provider can be picked (or no provider), and dependents can be rechecked.

when a file is deleted before the server starts, we can't parse it to extract its `providesModule`. if the file is using haste name reducers, where the module name is derived from the file name instead of `providesModule`, we know which module to dirty just from the "changes since mergebase" scm query.

if we're using saved state, the deleted provider is available when we parse the files changed since the saved state during init, but we ignore `changed_modules` in `init_from_saved_state`. since the file is also changed since mergebase, we also "recheck" it right after init and have to guess from the filename since it's already gone from shared mem. that isn't fixed in this diff.

so lazy mode correctness relies on using git or hg, and using haste name reducers. with some additional work, I think we could also make it correct without haste reducers, when using saved state.

Changelog: [fix] Fix an issue with lazy mode where deleting a file before the server starts did not check files that were depending on the deleted file.

Reviewed By: samwgoldman

Differential Revision: D42941765

fbshipit-source-id: 827697dda0fa9c258cefcdd74d5ae6429f38e5cf