[lti] Add `Comp_MaybeT` hint eval op for LHS of logical expressions
Summary:
We are exploring whether we should completely ignore the return hint in implicit instantiation if using the hint can cause an error. This approach looks promising, but it reveals an existing problem:

```
const safeParse = <T>(json: ?string): ?T => null;
const foo: Array<string> = safeParse('') ?? [];
```

This will fail because our current hint on `safeParse('')` is `Array<string>`. Then a subtyping of `Maybe<tvar> ~> Array<string>` failed, so we completely ignore the information provided by the return hint.

Whether we should go with the new approach is still up-for-debate, but we can at least fix this example by giving it a correct hint with a wrapped MaybeT.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D42970571

fbshipit-source-id: 540483d7203d7b16588472122b4c3b85aaf2d4dc