improve hint computation on calls to overloaded functions with union args
Summary:
Thanks to the special behavior of `PreprocessKitT` with union-like types, when determining
the winning overload, it is possible that we get multiple candidates from different branches
of the union. Before, we would always deem such a result as invalid as "overloads from sibling
branches are unlikely to lead to a helpful hint".

With fully resolved environments, code like
```
declare function overload(string, (string) => void): void;
declare function overload(number, (number) => void): void;
declare var foo: ?string;
overload(foo || "", (s) => {});
```
will cause the input argument to the call to be a union. In this case, both members of the union
will be strings, so they are going to agree to the final result of overload resolution.

So, we relax the initial condition to account for this case, and accept the overload for which
all members of the union agree on.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D42980361

fbshipit-source-id: f4cfa301d8f4a695163eda05ed184984a1911731