[IDE] Start replacing typechecker hooks in getDefUtils with typed ast
Summary:
This is the beginning of work to improve the performance of the document highlight command. The plan is to do this by using the cached typed ast rather than relying on typechecker hooks.

This diff focuses on removing the typechecker hooks that look for properties and replacing this behavior with a "searcher" that walks through the AST looking for property nodes. I also delete the now unused typechecker hooks. The only typechecker hook left in `getDefUtils` is the `obj_to_obj_hook`. This hook is special as we cannot replicate the behavior in a typed ast traversal. I'll target it in a stacked diff.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D42653145

fbshipit-source-id: 1bf694123d82f5aee3848c777d16d244ab6e99ea