[lti] Make instantiation cache a no-op under LTI
Summary:
## Instantiation Cache Breaks under LTI

Consider the following program:

```
declare var mixedArray: Array<string> | Array<number>;
const list = mixedArray.slice();
list.forEach(elem => {});
```

This is currently not well behaving under LTI. If you hover on the type for `list`, you will get `Array<string>`, which is clearly wrong. When you commented out the `list.forEach` line, hovering on `list` now shows the correct type. So what's happening here?

It turns out that, the instantiation cache is to blame. This cache was introduced in classic mode to solve the non-termination problem with implicit instantiation. This cache uses a bunch of reasons as cache key, and it will cache the result for the "same instantiation", and no subsequent cache hit, do a unification to ensure that the "same instantiation" will infer the same type.

A big issue is that, the cache's judgment on "same instantiation" is theoretically wrong, because a list of reasons contain no knowledge of the context and call stack the instantiation is happening, but just practically good enough in most cases under the classic inference. However, the unification scheme all breaks down under LTI.

Let's consider this example again:

```
declare var mixedArray: Array<string> | Array<number>;
const list = mixedArray.slice();
list.forEach(elem => {});
```

After type checking `const list = mixedArray.slice();`, the type of `list` is correctly inferred to be `Array<string> | Array<number>`. This is a union type of two `TypeApp`s. Then in the next line, when we are going to do `MethodT` check, `TypeApp ~> MethodT` will first trigger a specialization `PolyT(ArrayT(T), T) ~> SpecializeT(string)`. This specialization is also an instantiation, which will be cached as `T=string`. Later when we are doing the same for `Array<number>`, we will unfortunately hit the cache, and we will be forced to do `T=string=number`, which causes an error.

(This issue not only cause these confusions, but it's also a soundness problem. e.g. replacing `string` with `empty` in this example, then all bad things can happen.)

Under the classic inference, we will also hit the same problem, but it's working as expected. because tvars are unresolved, so we are actually unifying `T = OpenT([string], []) = OpenT([number], [])`, which is fine and does give us what we want.

## Safe to Kill the Cache

I attempted multiple fixes, but none of them are successful. I finally tried to kill the cache under LTI and it works. Remember that the cache is there to prevent non-termination issues instead of for perf. Such non-termination issues are no longer possible under LTI, because the lifetime of a mutable tvar is limited to the site of instantiation, and the instantiation will have fully resolved types as input. Therefore, it's impossible for this tvar to leak and cause some circularities which can lead to non-termination.

The unit tests under LTI and internal repo tests corroborated the claim above.

Reviewed By: jbrown215

Differential Revision: D42983248

fbshipit-source-id: f988497506293ef304ffc54d8c76dd04ca2d40d5