Fix to predicate function expressions
Summary:
Synthesizable predicate function expressions didn't act as though they were synthesizable--instead they acted as unsynthesizable functions. This led to LTI ordering violations, because LTI would consider the body of the function to be checkable after the function sig. This diff ensures that FunctionValues that are synthesizable predicates are correctly handled.

Closes #9004

Reviewed By: SamChou19815

Differential Revision: D43008282

fbshipit-source-id: 112f086b9c7eb2bf7a7de97ef1d055cade230456