Fix `this` in function class properties
Summary:
Currently a function class property without an annotation (e.g. `class C { foo = function() { this } }`) will treat `this` as empty. This is because the `this` type for the function will be resolved during resolution of the overall class type by calling `Func_sig.functiontype`, but the type that gets passed in is an open tvar  without bounds that gets resolved to empty. This diff changes that behavior by threading through the this_t for the class and using that as the this_t for the function. (This was the behavior prior to D38754956 (https://github.com/facebook/flow/commit/1abcd15ec57e9f56af8f35e84d25ff9acfce3007), cc SamChou19815 in case this change conflicts with something that you're aware of)

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D43007290

fbshipit-source-id: a89636bd3e83a9d4a9d1cc1be9b862aef598c151