Remove unused lint filtering
Summary:
This code attempts to filter non-active lints by checking the severity cover at
the point where we find a lint.

However, the severity cover is set by `scan_for_suppressions` which is not
called until after inference, so this filtering never actually fires.

This unused filtering was originally added to avoid accumulating a huge number
of lint errors in the server process. After this code was added, we added some
code which filters lints before sending errors back to the server, so the
original problem is solved without this code.

This filter is the only reason we need to look at `severity_cover` during
checking. Removing it enables some follow-up simplifications.

Reviewed By: mroch

Differential Revision: D43005877

fbshipit-source-id: 87c00c26007e012286c26638bcc4c029b6e16f95