fix "index out of bounds" exception in Fuzzy_score
Summary:
the tables used by `Fuzzy_score` have 128 rows and columns, to support patterns and words up to 128 characters respectively. but row 0 and column 0 are base cases and are filled in starting from (1, 1), so scores of the 128th character currently overflow and cause an "index out of bounds" exception.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D43047494

fbshipit-source-id: c3024cded1dde1e595495e3624e35393fd960216