[parser] Parse and error on JSX type arguments
Summary:
TS supports JSX type arguments. I think it would be nice to parse and error on them, so users familiar with TS would get a nice error message rather than a bunch of odd parser errors if they try to use the feature in Flow. (Also if for some reason we want to make use of them in the future, we'll have the option to do so.)

Closes #7675

Changelog: [parser] Parse and error on JSX type arguments

Reviewed By: jbrown215

Differential Revision: D42763585

fbshipit-source-id: 7f7d87f9b05e4d9a6401623c4bdc02fc356c5fc7