only return "strong" autoimport matches
Summary:
vscode (by default) only shows "strong" completion matches: the first character of the needle has to match the first character of the haystack, or just after a word boundary. A word boundary is a transition from lowercase to uppercase (the B in fooBar), or following punctuation (the b in foo_bar).

we previously returned weak matches, which has two impacts: (1) if "server controlled ordering" is on, then vscode's filtering doesn't work and we'd show weak results when we shouldn't; and (2) without server controlled ordering, we'd return 100 results and then vscode would filter a bunch out, so we wouldn't show as many results as we should.

this diff enables filtering to only strong matches server-side.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D42953069

fbshipit-source-id: 6b69c7b8e03ebccf33b9bd420f653c8168d81ad7