add first_match_can_be_weak to Fuzzy_path
Summary:
VS Code's fuzzy scoring has the concept of a "strong" match, where the first character of the needle has to match the first character of the haystack, or just after a word boundary. A word boundary is a transition from lowercase to uppercase (the B in fooBar), or following punctuation (the b in foo_bar).

In VS Code, a strong match is required by default, though that can be disabled with the editor.suggest.matchOnWordStartOnly setting. Since defaults normally should be falsy, the option in code is `first_match_can_be_weak`.

I replicated that in `Fuzzy_path`. But to maintain previous behavior of not caring about strong matches, our default is `true`.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D42952732

fbshipit-source-id: 2eb48e71d3144510939340378fa577031c884ebf