Add map callback to get_resolved_modules_unsafe
Summary:
Before this diff, this function always returns a map of Modulename.t values.
However, in many cases we immediately use that Modulename.t to find a dependency
addr.

This diff adds a map callback, so the caller can avoid this extra work and just
get a dependency addr directly. For now, this diff preserves existing behavior.

Reviewed By: mroch

Differential Revision: D42878808

fbshipit-source-id: 1c848a66982db38546cbfac3a3191fb1cb963273