Remember phantom dependency addrs during module resolution
Summary:
Similar to the previous diff, but for phantom dependencies instead of resolved
modules.

It's possible for a haste module / file to exist without a provider file. In
this case, the module fails to resolve because it doesn't "exist" but we already
have a tracking record for it.

By remembering the addresses, we can avoid redundant hash table lookups during
resolve requires.

Reviewed By: mroch

Differential Revision: D42878806

fbshipit-source-id: 682a8959af85743b6e8a54ad0d93487cc0ec0931