Remember dependency addrs during module resolution
Summary:
We know a module exists by looking in the hash table. If we find a haste module
or file that has a provider, then the module "exists" otherwise it doesn't.

This diff returns the address of the dependency to the caller instead of just
the Modulename.t. This does two nice things:

1. Many callers just want the provider, and can now call `get_provider` without
a redundant hash table lookup.

2. For resolve requires, we don't need to do any additional hash table lookups
for resolved modules. We get an array of addresses, which is exactly what we
want.

Reviewed By: mroch

Differential Revision: D42878797

fbshipit-source-id: da2b83e9d31370593d0b8e15288151e11c620f78