Pass in resolved modules and phanto dependencies separately
Summary:
We pass in "resolved requires" in two places: (1) during the resolve requires
step in non-saved-state init and recheck and (2) when loading saved state.

In this stack, the signature for (1) is going to change and the type of resolved
requires will be different from (2).

This diff removes `resolved_requires` type and changes the API to take the two
fields separately.

Reviewed By: mroch

Differential Revision: D42878804

fbshipit-source-id: 061b47685b003ceae7fdb817f8148443176dff58