Allocate resolved requires separately
Summary:
This diff is where the preliminary diffs start to come together.

In this diff, we add resolved requires in two allocations:

First, for each resolved module and phantom dependency, we find the dependency
addr for each, creating one if necessary.

Second, we compare any old resolved requires if necessary. Because all
dependency addresses are resolved at this point, we can compare addresses
instead of module names, which is more efficient.

Similarly, when updating reverse dependencies, we can partition based on
addresses instead of comparing module names, which is more efficient.

Reviewed By: mroch

Differential Revision: D42878796

fbshipit-source-id: 5857ab98b89a921e7fe71334fd4b186b4b0089ce