Add map callback to read_resolved_requires
Summary:
Before this diff, `read_resolved_requires` always returned a `resolved_module array`, but in many
cases we just want the `resolved_module_addr` instead, which is much cheaper. This diff breaks up
the current logic into two functions, composable though a callback, and updates existing callers.

This is used later in the stack (along with a few more similar changes) to avoid unnecessary work
when traversing the dependency graph.

Reviewed By: mroch

Differential Revision: D42878807

fbshipit-source-id: 1571a716ed7232d2e79f3c8516bbfe2fa73a9ebf