Use dependency addrs directly during check
Summary:
Earlier in this stack, we were able to use the dependency addrs from
`get_resolved_modules_unsafe` and `Module_js.imported_module` to call into
`get_provider` efficiently: we can avoid allocating Modulename.t values and
unnecessary hash table lookups.

For check, this is a little more complicated because the code is parameterized
by a "reader" module, and thus generic. It does not depend directly on the
parsing heap, or even shared memory at all.

This diff updates the "reader" module signature, allowing the "heap reader"
implementation to be more efficient.

Reviewed By: mroch

Differential Revision: D42878802

fbshipit-source-id: ecc12c967d5555698f4f4921b76df31240d1b649