[IDE] use a cacheable obj_to_obj_hook in GetDefUtils
Summary:
As mentioned in the last diff, the `obj_to_obj` hook cannot be replaced directly by an AST traversal. It keeps track of mapping object literals to type definitions, potentially through a type alias. The plan to improve performance is to run the `obj_to_obj_hook` on every typecheck run for the IDE (the same typechecked that produces the cached typed AST). This diff rewrites the hook to take no arguments and just produce a map from locations to sets of properties. This map is then used in `def_info_of_typecheck_results` to look up the definition prop location (which used to be computed from the type).

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D42653144

fbshipit-source-id: 93774f7a24fc54fbb330fca20690d27a008a973e