[lti] Add AnyErrorHint as another base hint node for name_def
Summary:
This is different from `Hint_Placeholder` in that an any type as hint is expected due to errors, but because we are too lazy to support a feature (like placeholder hint for destructuring assignment that I will fix in the stack). This allows us to kill Hint_Placeholder eventually.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D43106152

fbshipit-source-id: 9b3ef7ae3b0000d38a11e9884eb9806c78313ab8