Make `unused-promise` enable both sync and async unused promise lints
Summary:
To ease upgrading, we'll have one version with both `unused-promise-in-async-scope` and `unused-promise-in-sync-scope`. `unused-promise` will enable both. Later, we can remove the individual rules.

Changelog: [feature] Added an `unused-promise-in-sync-scope` lint which will detect unused promises in sync scopes (like `unused-promise-in-async-scope` does for async scopes). Enabling `unused-promise` will enable both of these rules. The individual rules will be removed in a future release.

Reviewed By: gkz

Differential Revision: D43158888

fbshipit-source-id: d7364f22a4c9f57e2d7097dba8245855765ab189