Preserve variance through `$Partial`
Summary:
Changelog: [errors] The type utility `$Partial` will now preserve the variance of the input object type's properties. For example, if the the input was a read-only object, now that output will also be a read-only object (with optional properties).

`$Shape` preserves variance, so with this `$Partial` will be fully better than `$Shape` in every way.

Reviewed By: SamChou19815

Differential Revision: D43112621

fbshipit-source-id: 8aab62686210b78830cd246a9c0f3af01aa3c3bd